Refactoring code, updated readme
`pi.jl` broken down into several manageable chunks.
